DRY DeepCopy Visitor.
This commit updates the DeepCopy Visitor class to remove duplicate
code. A new method `deep_copy_attrs` is introduced which accepts a
variable number of arguments each referencing a property defined on
the target node. Each child of the property is de-referenced by
calling `deep_copy` on it. The result is assigned back to the
property being enumarated.