First working draft, implement {Buffer,Flush}Node/s in Perform.
This commit implemented the @buffer and @flush directives in the
Perform Visitor.

The visit_children(..) method is updated to support tracking of the
parent so bubble(..) can work as expected. When we replace the tree
in the buffer, we request the contents to be wrappen in an array
[..] instead of using the Array(..) cast. The latter would flatten
the tree resulting in duplicate nodes.

Many tests were added mostly to validate functionality and to
support the code updates, not to demonstrate usefulness.